refactor(renderer): extract composer/InputArea cluster
Third extraction wave for BET-47. Moves the bottom composer and its
helpers (SessionToolbar, AttachmentStrip, TypeaheadPopup, MicButton,
InputArea) into InputArea.tsx, and hoists the shared composer types
(Attachment, AgentMention, TypeaheadState, TypeaheadRow) into chatShared.

ChatPanel.tsx drops from ~4960 to ~4190 LoC. typecheck passes.
